Add support for Ruby 2.1
Make use of the new TracePoint API. We cannot rely on our own C hooks
anymore, because in Ruby 2.1 they've removed the information about
raised exceptions from the thread. They needed to fix a bug in
`set_trace_func`[1][2]. Namely, this commit[3] breaks Interception.

I couldn't think of a way to patch Ruby in order to make it possible to
use the same mechanism we utilised before and keep their fix alive. I
must admit it's been a *really long* dive in Ruby internals and I'm not
very familiar with them. Anyway, I think TracePoint is the way to go,
since it probably will be forever with Ruby from v2.0.0.

[1]: https://bugs.ruby-lang.org/issues/9168
[2]: https://bugs.ruby-lang.org/issues/9088
[3]: https://github.com/ruby/ruby/commit/cf5ef3db5e33195ef8cbf5288456f60c6d90f75c